feat(restaurant) : 식당 삭제 API 인증 처리 및 주인 검증로직 추가

+++ src/main/java/com/example/fooddelivery/restaurant/service/RestaurantService.java
@@ -106,7 +106,9 @@ public class RestaurantService {
     }
 
     @Transactional
-    public void deleteRestaurant(Long restaurantId) {
+    public void deleteRestaurant(String identifier, Long restaurantId) {
+        Restaurant restaurant = findRestaurantById(restaurantId);
+        validateOwner(identifier, restaurant);
         restaurantRepository.deleteById(restaurantId);
     }
 }
